pyproject.toml
[project]
name = "everlight-api"
version = "0.1.0"
description = "REST API Service for Everlight CRUD and Integrations"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "firebase-admin>=6.9.0",
    "pgvector>=0.4.1",
    "psycopg[binary,pool]>=3.2.9",
    "sqlalchemy>=2.0.41",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.23.0",
    "notion-client>=2.4.0",
    "testing-postgresql>=1.3.0",
    "pytest>=8.4.1",
]
